Move exif.h to base/
