+++ lib/jxl/enc_fields.cc
@@ -5,6 +5,8 @@
 
 #include "lib/jxl/enc_fields.h"
 
+#include <cinttypes>
+
 #include "lib/jxl/enc_aux_out.h"
 #include "lib/jxl/fields.h"
