+++ lib/jxl/fields.h
@@ -8,15 +8,12 @@
 
 // Forward/backward-compatible 'bundles' with auto-serialized 'fields'.
 
-#include <inttypes.h>
-#include <stddef.h>
-#include <stdint.h>
-#include <stdlib.h>
-#include <string.h>
-
-#include <cinttypes>
 #include <cmath>  // abs
 #include <cstdarg>
+#include <cstddef>
+#include <cstdint>
+#include <cstdlib>
+#include <cstring>
 
 #include "lib/jxl/base/bits.h"
 #include "lib/jxl/base/compiler_specific.h"
