+++ lib/jxl/modular/transform/transform.cc
@@ -5,6 +5,8 @@
 
 #include "lib/jxl/modular/transform/transform.h"
 
+#include <cinttypes>
+
 #include "lib/jxl/base/printf_macros.h"
 #include "lib/jxl/fields.h"
 #include "lib/jxl/modular/modular_image.h"
