+++ lib/jxl_lists.cmake
@@ -14,6 +14,7 @@ set(JPEGXL_INTERNAL_BASE_SOURCES
   jxl/base/common.h
   jxl/base/compiler_specific.h
   jxl/base/data_parallel.h
+  jxl/base/exif.h
   jxl/base/fast_math-inl.h
   jxl/base/float.h
   jxl/base/iaca.h
@@ -185,7 +186,6 @@ set(JPEGXL_INTERNAL_DEC_SOURCES
   jxl/entropy_coder.h
   jxl/epf.cc
   jxl/epf.h
-  jxl/exif.h
   jxl/fast_dct-inl.h
   jxl/fast_dct.cc
   jxl/fast_dct.h
