+++ tools/cjxl_main.cc
@@ -43,11 +43,11 @@
 #include "lib/extras/time.h"
 #include "lib/jxl/base/c_callback_support.h"
 #include "lib/jxl/base/common.h"
+#include "lib/jxl/base/exif.h"
 #include "lib/jxl/base/override.h"
 #include "lib/jxl/base/printf_macros.h"
 #include "lib/jxl/base/span.h"
 #include "lib/jxl/base/status.h"
-#include "lib/jxl/exif.h"
 #include "tools/args.h"
 #include "tools/cmdline.h"
 #include "tools/codec_config.h"
